Revert "Implemented fix from early issue #16.  Finally decided to pull the"
This reverts commit 476097496a35a2caa1a586be004dc1a511ba7e9f.

Reverting the change from issue #16.  After some discussion, it has
been decided that it is up to the user to implement the pull-up and
pull-down modes correctly by setting the output enable and driving
the output to the appropriate value.  Note that this should be well
documented, if by nothing else than a validation testbench that
excercises a user pull-up and pull-down input mode.